Add typed DateTimeError; change u64 -> i64 for timestamp and constants; address other code review comments

+++ radix-engine-interface/src/time/constants.rs
@@ -1,3 +1,3 @@
-pub const SECONDS_IN_A_MINUTE: u32 = 60;
-pub const SECONDS_IN_AN_HOUR: u32 = 3_600;
-pub const SECONDS_IN_A_DAY: u32 = 86_400;
+pub const SECONDS_IN_A_MINUTE: i64 = 60;
+pub const SECONDS_IN_AN_HOUR: i64 = 3_600;
+pub const SECONDS_IN_A_DAY: i64 = 86_400;
